Move to fboss/util directory
Summary: To avoid having to rewrite a lot of common code in FixMyFboss|BMC we can publish fixmyfboss to a public part of the fboss repo. In openbmc we can use SRC_URI to import the code (excluding the modules/ directory which contains fboss-specific checks). Then we will probably need to sed the files to fix the import statements but once that is done we will have free syncing between fixmyfboss and fixmybmc. Modules in openbmc will be unchanged

Reviewed By: kimdo8736

Differential Revision: D70575272

fbshipit-source-id: 6f72e8c4923f850cf514af0998a7adedbfdd1326