Update system port SHEL destination enable for switch SHEL config update
Summary:
As titled. Currently we enable RCY port's SHEL enable if SHEL switch setting is present.

In the case of disabling SHEL, switch setting will be removed. We need to update all RCY ports in the cluster to disable SHEL destination, effectively disabling SHEL.

This is to handle the logic in ApplyThriftConfig.

Reviewed By: jasmeetbagga

Differential Revision:
D70151987

Privacy Context Container: L1125642

fbshipit-source-id: 8f391326e806cc86f30a44f9c6499ffc9c0bb4b2