Prevent unbounded pending updates queue build up
Summary:
Updates to be published to FSDB are enqueued by FsdbSyncManager to its internal
managed storage (CowStorageMgr). At present these updates are queued in
CowStorageMgr as non-coalescing updates, which can result in unbounded build up due
to back-pressure from sequential 1-at-a-time handling of non-coalescing updates in
stress scenarios.

This change is to enqueue the pending updates as non-coalescing so that every time
CowStorageMgr::handlePendingUpdate() dequeues all the pending updates queued up
and a single coalesced update is handled by FsdbSyncManager for publishing to fsdb.
Therefore with this change, when there are multiple pending updates queued the cost
of computing the OperDelta and publishing to FSDB is reduced to a single such operation
to prevent back-pressure on the pending updates queue.

We will measure max queue length in CowStorageMgr with max-scale + max-churn to
determine the threshold for max pending updates in the CowStorageMgr queue, and
in next diff export quantile stats for the max_pending_updates queue watermark for
monitoring.

Reviewed By: jasmeetbagga

Differential Revision: D70566619

fbshipit-source-id: 9cd8bfe24dc88821b9b5b65cc90358ca58b4379c