Add support for AEC cables in QSFP Service
Summary:
Support AEC cables in the QSFP Service. With all the refactoring in the previous diffs, this is mostly just a matter of checking if the module is an active cable and then applying Active cable application codes etc.

Templatizing getMediaInterfaceCode which would be used for both SMF and Active Electrical Cables.

Reviewed By: harshitgulati18

Differential Revision: D71103830

fbshipit-source-id: 43e7025b62b4c22d861bcc80c34271213e9e3d34