Adjust PidLogic to not let integral slow down the response
Summary:
When the measurement (temperature) is below setPoint (expected levels)
for a prolonged perioed of time, the integral_ gain kept accumulating. When
the temperature suddenly increases over the setPoint, then this large integral_
will delay the response of the algorithm.  i.e., PWM of the fans will not be
increased quickly.

To address this, we do not let integral_ gain to grow indefinitely when the
temperature is below setPoint. This mirrors the adjustment made in the
oldAlgorithm for this use-case.

Reviewed By: mikechoifb

Differential Revision: D71509861

fbshipit-source-id: f6c94550c116cd8aed74332a6030e7df42f12984