Modify switchReachability to use fabricPortGroupId as 0 for empty port groups
Summary:
As of now, we just assign an unused ID as the fabricPortGroup ID for any new port
group. For clients like BGP using dsfSwitchReachability information in FSDB, the only
information needed is if a switch is reachable or not. For this, clients need to now get
the switchID to fabricPortGroupId, then do another lookup in fabricPortGroupMap to
see if an entry exisits and if it exists, check if the group is empty or not. To make it
easier for clients, we are looking at using a fixed fabricPortGroupId of 0 as the empty
one, which means if switchId maps to fabricPortGroupId of 0, then it means the switch
is not reachable and if its non-zero, it means the switch is reachable, helps avoid doing
one more lookup, at the same time, retaining all the information as before in the
dsfSwitchReachability.

Reviewed By: jasmeetbagga

Differential Revision: D71229867

fbshipit-source-id: f22993dd47e81699428f287532a24c1798c4085f