Provide clear error message when not run as root
Summary: Provide clear error message when not run as root.This diff is related to the FBOSS platform weutil. The code changes include adding a new function called "check_root" which checks if the user running the program has root privileges. If the user is not root, the program will exit with an error message.

Reviewed By: somasun

Differential Revision: D71638855

fbshipit-source-id: 7cce26e76a6bdee4aec0f042f79235c5a88169b3