Remove UT that's blocking neighbors
Summary:
This is towards deprecating the use of `setNeighborsToBlock`. We updated to use 'setMacAddrsToBlock' more than 2yrs ago, after realizing multiple IPs could be tied to the same MAC address.

To avoid the API affecting lingering usage, we've verified
1. No more reference of `setNeighborsToBlock` in our code base.
2. No call in the acl check for this API.

For this diff: Remove UT's that's reliant on logic of processing blocked neighbors.

Reviewed By: shri-khare

Differential Revision: D71594349

fbshipit-source-id: 8daffe9391f0e5f76fe53e2e01a60e39dedc60e9