Stop processing blocked neighbor in lookup class updater
Summary:
This is towards deprecating the use of `setNeighborsToBlock`. We updated to use 'setMacAddrsToBlock' more than 2yrs ago, after realizing multiple IPs could be tied to the same MAC address.

To avoid the API affecting lingering usage, we've verified
1. No more reference of `setNeighborsToBlock` in our code base.
2. No call in the acl check for this API.

For this diff: No need to process blocked neighbor in lookup class updater.

Reviewed By: shri-khare

Differential Revision: D71594577

fbshipit-source-id: b55a900ae2d07848e2e1461b826993e7b52f46fd