Remove getter/setter of blocked neighbors in switch settings
Summary:
This is towards deprecating the use of `setNeighborsToBlock`. We updated to use 'setMacAddrsToBlock' more than 2yrs ago, after realizing multiple IPs could be tied to the same MAC address.

To avoid the API affecting lingering usage, we've verified
1. No more reference of `setNeighborsToBlock` in our code base.
2. No call in the acl check for this API.

For this diff: Remove getter/setter of blocked neighbors in switch settings.

Reviewed By: jasmeetbagga

Differential Revision: D71595732

fbshipit-source-id: 9b366d2e98c8402fb7b9740f48d29612bfa5fd80