Track ports with recently updated config
Summary:
Create a tracker for ports with recently updated configurations.

FabricConnectivityManager already has a method to calculate connectivity
changes and generate a delta for processing, but is not able to generate a
delta if configuration changes from the incorrect to the correct cabling.
This tracker allows SaiSwitch to force-generate a connectivity delta in this
scenario.

The usage of this method in SaiSwitch will be added in a future diff in stack.

Reviewed By: daiwei1983

Differential Revision: D71497226

fbshipit-source-id: 226a3424e07c6a72443e302983dc6539583d0c9c