Generate port delta if config changes
Summary:
Force the generation of a connectivity delta if a port config is
updated. This addresses the edge case where a port config can be applied to
change to correct connectivity, but no connectivity delta is generated so the
port errors are not updated.

Hardware test will be in following diff.

Reviewed By: zechengh09

Differential Revision: D71770773

fbshipit-source-id: e0b112d4bf56ccbcc50a886dff26b922cd4abfd6