keep vlan id optional for chenab along side voq switches
Summary:
port rif are traffic ports in chenab. ignore chenab in considering vlan processing in rx packet handling.

note - this needs to be cleaned up in a way that rif type (or port) should govern vlan presence in  packet and not asic type or switch type. planning to stack diffs later as scope is general cleanup and not chenab specific

Reviewed By: vasantspatil

Differential Revision:
D71698137

Privacy Context Container: L1125642

fbshipit-source-id: 4f81dda94e6781b5164c57674df2b2daefc44005