fix packet buffer unit size
Summary:
Updates the buffer cell size for Chenab, so that the calculations in tests calculate the correct number of packets.

Adapts the test toloop packets out of the front panel port rather than sending directly from the CPU as pipeline lookup.  We noticed an issue (under investigation) with ECN marking for pipeline lookup, but front panel traffic seemed like a more realistic test.

Differential Revision: D71982286

fbshipit-source-id: 9f8b17327c53f05cd0533d2a5fc9fa369d7a2cd4