Change raw string delimiter to allow parentheses
Summary:
D71938274 is a fw_util config change from Arista which introduces a chip named "GD25Q16(B)". This causes ConfigLib to fail since the delimiter of the raw string `R({})` is just a parentheses. When the ")" is found in the config, the raw string is closed.

This diff modifies the delimiter to `EOFEOF({})EOFEOF` which will probably never be encountered in a config file

Reviewed By: somasun

Differential Revision: D71990001

fbshipit-source-id: 3d35816d62ed5673eb96bf4f9279a0610367c217