Add num-jobs flag to docker-build.py
Summary:
docker-build.py makes a call to getdeps.py which contains control over parallelism. This diff will add a flag "num-jobs" to docker-build.py to more easily control the behavior of getdeps.py

this will help us in the future in the case that docker-build.py is causing memory to fill up during builds.

Differential Revision: D72008533