Generate soc file to overwrite RTP GPD setting in R3 to avoid S503928
Summary:
Generate a soc file which can set the `RTP_GRACEFUL_POWER_DOWN_CONFIGURATION`
register to 0 after SAI/SDK init is complete. This will ensure that on undrain, RTP GPD
feature will not interfere with traffic forwarding.

As this is needed on R3 devices in 2-stage, making sure that only one of the 2 hw
agents will create the file. The file generation is in pre-init and the file will be consumed
post init and controlled via asic specific config in D71998682.

Reviewed By: jasmeetbagga

Differential Revision: D71986447

fbshipit-source-id: 8611bc253f050cc768e16a0837abd36b2615100f