Remove nested transceivers_.rlock() in refreshTransceivers()
Summary:
Introduction of `resetTransceivers()` Thrift API introduced a potential deadlock scenario to `qsfp_service`. These deadlocks occur when `refreshTransceiver` is running and is waiting for each module to finish its `futureRefresh` call. Some notes:

- We use `folly`'s synchronization primitives to control access to `transceivers_` map.
- `refreshTransceivers()` thread runs every 5 seconds.
- `refreshTransceivers()` acquires `transceivers_.rlock` at the beginning to determine which transceivers to initiate `futureRefresh` on.
- `refreshTransceivers()` holds this lock until all transceivers have completed their `futureRefresh` call.
- `futureRefresh()` function can require the execution of `verifyEepromChecksums()` and `setDiagsCapability()`, both of which require an `rlock` on `transceivers_` (this means our code introduces a nested `rlock` scenario which should be harmless most of the time).
- When `resetTransceiver()` Thrift call is made, which requests`transceivers_.wlock`, in between the time `refreshTransceivers()` call starts and ends, a deadlock arises.

This all happens due to folly's implementation of `SharedMutex`, outlined in a comment in https://fburl.com/workplace/9mel8lt3 – "SharedMutex has write priority, meaning that if a writer is waiting to acquire the mutex, no new readers are allowed until the writer succeeds in acquiring and then releasing the mutex." This is to avoid writer starvation.

In our scenario, `resetTransceiver` can't proceed forward with acquiring `wlock` until `refreshTransceiver` releases its `rlock`. However, due to this `SharedMutex` rule, each module's `futureRefresh` / `verifyEepromChecksums` call can't acquire a new `rlock` until `resetTransceiver` acquires its `wlock`, which is impossible. Thus we reach a deadlock.

This diff uses folly provided `unsafeGetUnlocked()` function to directly access the transceiver level data in `verifyEepromChecksums()` and `setDiagsCapability()`, two functions that are only called when `refreshTransceivers()` is running and holds the `rlock`. These functions should only be used in a locked context, and to enforce this, we added a new function `isTransceiversMapLocked()` which verifies some thread holds an rlock / wlock on the map at this time.

--

We temporarily opted for this solution instead of giving each `TransceiverStateMachine` a `Transceiver*`. `transceivers_` right now is a map of TcvrIDs to `unique_ptr<Transceiver>` – changing this to a `shared_ptr` could make managing the lifecycle of a Transceiver more difficult and require more code changes – we can consider approaching this design later.

Reviewed By: harshitgulati18, xanabani

Differential Revision: D71160485

fbshipit-source-id: 80a189642eb48dd12d4a071bc7994751c5034b94