Publish subscribe_latency_ms as per-client metric
Summary:
FSDB SLI for subscribe_latency_ms was added in D70648880 as a per-subscription
metric. This requires use of per-subscription unique keys to export the metric
to ODS, resulting in too many time series when composing ODS queries for per-DC
monitoring.
For FSDB SLO monitoring exporting the metric on a per-subscriber level
suffices. So this change migrates to using metric key
"FsdbClient.{state|stats}.subscribe_latency_ms".

Also, for any clients that may be interested in exporting per-subscription
granularity metrics, adding a SubscriptionOption flag that allows continuing to
export the metric with current scheme.

Differential Revision: D72187883

fbshipit-source-id: 420625f77590aac2a7522819cadeddfeb00d1086