Add num-jobs flag to docker-build.py (#370)
Summary:
Pull Request resolved: https://github.com/facebook/fboss/pull/370

docker-build.py makes a call to getdeps.py which contains control over parallelism. This diff will add a flag "num-jobs" to docker-build.py to more easily control the behavior of getdeps.py

this will help us in the future in the case that docker-build.py is causing memory to fill up during builds.

Reviewed By: srikrishnagopu

Differential Revision: D72008533

fbshipit-source-id: badde7c539f4a9b9af9aa8a1f3510bd37e6fda08