Add verifyWredThreshold
Summary:
Migrate HwTest verifyWredThreshold test to AgentHwTest

This diff adds a new function `verifyWredThreshold` to the `AgentAqmTests.cpp` file to ensure that the number of dropped packets is as expected. The function is used to verify the number of dropped packets due to WRED threshold. The function takes in the `AqmTestStats` objects before and after the test and the expected number of dropped packets. It then calculates the delta in dropped packets and checks if it is within an acceptable error percentage.

Reviewed By: maxwindiff

Differential Revision: D72172033

fbshipit-source-id: 8983ba77e9f36ac0d1f6401bca92053da7a3f81f