Strengthen AQM tests queue buildup logic
Summary:
AgentAqmTest verifyWredThreshold test fails on Tomahawk1 because it may take a little while for the software switch stats to update to detect any leaked packets from residual credits left on the switch. Fix this by calling a stats update before sending packets to build up the queue and a wait for the number of out packets to be positive.

Add a check to wait for the number of out packets to be positive or until the number of retry iterations has reached the maximum.

Include some function refactoring and use signed instead of unsigned integers in arithmetic for overflow safety.

Reviewed By: nivinl

Differential Revision: D72468131

fbshipit-source-id: 4f2384a3ee12841781276f6bb23bb2eb41268c23