Show raw lbw value when printing ext communities
Summary: For NSF we cannot interpret the ext community value as float coz it won't make sense. Right now we don't rely on any flag to indicate whether the lbw is encoded, so we print out the raw value (like we do with all other ext communities) along with the translated lbw.

Differential Revision: D72696686

fbshipit-source-id: 59df8a8e99da9967af82886de35b79c53417cef1