Publish dummy data from agent to FSDB for metric benchmarking
Summary: Publish dummy data in Agent Churn test after all the churn is done to signal FSDB that churn is complete. The subscription callback will then wait for the publish to be complete. It then notifies once it receives the signal. The publish, subscribe and latency are already being measure in FSDB and also exported to ODS. These can be used to monitor the threshold.

Differential Revision: D72411244

fbshipit-source-id: d9e016f5404e89f50479b71e6183172df16076d4