Fix queue latency VOQ test
Summary:
The test disables tx. It wants to verify Voq queue latency count exceeds the max threshold set, as packets are stuck.

However, enabling the tx right after may reset the latency,  before the test gets to verify the out of bound latency. This makes the test flaky,

Make the test more deterministic by checking the latency while packets are stuck and not tx is enabled.

Reviewed By: nivinl

Differential Revision: D72773685

fbshipit-source-id: dbd66ef0bf15401db247a443f9c8e0a4a86851b7