Retain class ID if multiple neighbor entry with same mac
Summary:
As titled. Currently if two neighbor entries are with the same mac, removing the first neighbor entry will result in mac entry losing its classID. This is because our workflow of

1. Remove mac entry,
2. Re-add mac entry if any neighbor still have the same mac.

Then LookupClassUpdater will re-assign classID to the mac entry, assuming the removed neighbor entry is processed prior to mac entry.

Now we're moving to single update for neighbor + mac entry, leading to mac entry losing its classId when this happen.

To fix this, retain classId when we're pruning mac entries.

Reviewed By: shri-khare

Differential Revision: D71775338

fbshipit-source-id: 17d4f9c981981ae8bb27342b38dc1f8a5720feb3