Move AQM test config building to initialConfig
Summary: This is a partial revert of D71898415. Config changes should (often) be confined to `initialConfig` or `setup`. While testing `AgentAqmTest.verifyEcnTrafficNoDrop` test, calculations are performed to determine the number of packets to send and ECN markings to expect based on the config before the `setup` function. If a config change happens after, the calculations may be incorrect for the test. Move `addOlympicQueueConfig` to `initialConfig` and restore `isDualStage3Q2QQos` check to call `addNetworkAIQueueConfig`.

Reviewed By: maxwindiff

Differential Revision: D73061607

fbshipit-source-id: 0a21c331ea383494e6aeeebb9110720af5133ad1