Set LOSSLESS=1 for TH4/TH5 as a workaround
Summary:
Buffer headroom watermark doesn't work unless PGs are put in lossless mode. CS00012395772 will set this based on PFC state, for now we need to set this in yaml as a workaround.

Note that this fix only applies to TH4/TH5 since TH3 uses a different SOC property format.

Reviewed By: nivinl

Differential Revision: D73007136

fbshipit-source-id: 5fd2831828196d0af83fae6aa8ae8d69c5f11c20