Fix bug in delete route handling
Summary:
We cant cache iterator for prefix2NhopGroupInfo map. This will
increment the reference count for weak_ptr held in nhopGroupIdToInfo map.
Which means that even if this was the last prefix referencing this nhop group,
we would see a reference of 1 and not delete it.

Differential Revision:
D73294486

Privacy Context Container: L1125642

fbshipit-source-id: a37c3010105d846e6fe3ba2459c718c24fada615