Remove dated requirements
Summary: Some requirements stated that "latest version of X must be used" and also listed the actual latest version at the time of writing. Just the first part is required so we don't need to keep updating a specific version in this doc which is public knowledge every time it changes.

Reviewed By: somasun

Differential Revision:
D73387230

Privacy Context Container: L1125642

fbshipit-source-id: 57999e76983c331892255f095bfe53c07c7be21a