Add working directory check to run_test.py
Summary:
Adds a check at the beginning of the script to ensure that it is run from /opt/fboss.

The script has some assumptions built into it that require this (e.g. the default unsupported test and known bad test files are relative paths and therefore won't be found if the script is not run from the correct location).

Reviewed By: shiva-menta

Differential Revision: D73472338

fbshipit-source-id: 523a49e27ca099de108b591677b5418f463a54b8