fix exception handling in SetEncodedPathVisitorOperator
Summary:
In ThriftCOW, set of primitives are implemented with immutable
ThriftPrimitiveNode. I.e. individual members of the set of primitives are
immutable.
When a set_encoded operation on a set of primitives is involved, after modifying
the path, SetEncodedPathVisitor operator visits the immutable ThriftPrimitiveNode
and invokes a write operation on it. This results in exception.

This bug goes back to when SetEncodedPathVisitorOperator was introduced.
This handling can be corrected by safely ignoring the exception as the immutable
primitive node in set does not require any update to its contents.

Towards ensuring code coverage, adding UTs for operations on sets of
primitives in this diff with which the issue was reproduced and fix is verified.

Reviewed By: Linerd

Differential Revision: D73581840

fbshipit-source-id: 555b28c20d5f25179be24a13235ce5712074193a