fw_util: Add a member variable to FwUtilImpl.h class to help handle printing the version of the Firmware
Summary: This diff adds a member variable to the FwUtilImpl.h class to help handle printing the version of the firmware. The code changes include adding a new header file for the version handler and updating the class to include a unique pointer to the version handler. This will allow the FwUtilImpl class to handle printing the firmware version.

Reviewed By: Scott8440

Differential Revision: D73378277

fbshipit-source-id: b3d52f31ff98d3b7e9da1e27a7355fbb9888cd65