Change OSS Tool Default Directories to fboss/lib/platform_mapping_v2
Summary:
- Changing default input directory path to be the one we expect them to put their source files when making a PR.
- Changing default output directory path to be the one we expect them to put their generated JSONs when making a PR.

This diff introduces a change in workflow where we're expecting vendors (and FBOSS devs) to check-in generated files to fbcode as well. This makes it easy for us to run GitHub Actions checks (in future diff) to check these files against what we expect to be generated with vendor changes in PR (D72916745).

Reviewed By: paulcruz74

Differential Revision: D72900565

fbshipit-source-id: 62a20d9650e8194b2dd00a82fec04a9575cb8beb