Add Support for Auto-Generated Configo Sync Diff
Summary:
This diff adds support for auto-generating Configo sync diffs that are dependent on any fbcode changes to platform mapping code. This makes it easier for devs to keep track of which diffs they're responsible for managing and landing to keep cfgr in sync with fbcode platform mappings.

This sync job operates similarly to `thrift-www-sync` jobs.

Reviewed By: birdsoup

Differential Revision: D73073738

fbshipit-source-id: 4aa88ffaccc36e01ce23f56ca9e6efefdb998862