Add retries to wait for packet transmission in HwAqmTests
Summary:
HwAqmTest verifyEcnThreshold fails on buck mode `opt` while it's passing on `opt-asan` which is the default buck build mode for running and executing the tests.

The test is not waiting for packets transmission to be complete causing the test to prematurely proceed to the verification stage yielding errors.

Add a retry to wait for packet transmission to be complete by checking the loopback port packets counter.

Reviewed By: nivinl

Differential Revision: D73618497

fbshipit-source-id: 5a519739516857292041796edb315e686e660d97