Start publishing pfc relater counters using the HwPortStats based approach in multi/mono
Summary:
PFC deadlock related fb303 counters are not populated in multi-switch mode.

Even in mono-mode, these counters are populated in a non-standard way.

During a SAI callback `pfcDeadlockNotificationCallback()`

https://www.internalfb.com/code/fbsource/[b75186bc7c6488e19bb33f872518f32baf14bed1]/fbcode/fboss/agent/hw/sai/switch/SaiSwitch.cpp?lines=4531

 , SaiSwitch calls `pfcWatchdogStateChanged()` callback.

For mono mode, this callback is implemented but for multi it is not

https://www.internalfb.com/code/fbsource/[fcc7bee6e23779f88648fafb7d1d58d3540deacd]/fbcode/fboss/agent/mnpu/SplitAgentThriftSyncer.cpp?lines=222-226

This leads to AgentTrafficPfcWatchdogTest completely failing in multi-switch mode https://fburl.com/conveyor/ncfb1igt

This diff moves away form this callback based approach to simply updating the counters in `HwPortStats`. This approach works for both multi / mono and is simpler.

Reviewed By: zechengh09

Differential Revision: D73636408

fbshipit-source-id: 30950a1dd3180d86ba2f0d67ffd068bb0d9321da