Add a check in tx pkt path for sw exiting in ipv6 handler
Summary: On wedge100s, it has been observed that swswitch can crash while sending a tx packet when the switch is exiting. SwSwitch::sendL3Packet already has a check but it looks like on wedge100s, a race can happen when the check passes for tx thread and then the sw switch can start exit. Logs - P1794895870 and pdb backtrace P1798046206. This is seen very rarely (2 times till now) and had the same backtrace. i am adding an extra check in the tx routines. The other option is to use a lock between the threads which i want to avoid

Differential Revision: D73795508

fbshipit-source-id: e8f418646fe1eeeaab3786138493ccec6202256d