Fix the log for DNX "interrupt masked events" and change log level as this can happen in prod
Summary:
If a real interrupt masking issue happens in fleet, there will be a lot of msgs and we dont need that
except for debugging to understand whats the interrupt ID causing the issue. The counter should be
good enough for us to figure that something wrong is happening and once seen, we could lower the
debug level to understand the interrupt ID causing the issue. Another option to debug this is to look
at the below and look for one with a large interrupt count.
```
fboss_bcm_shell -c "show int"
```

Reviewed By: jasmeetbagga

Differential Revision: D74037136

fbshipit-source-id: 250ec4b1ae536354325d59588d21c64b0b3b43bb