Fix test ensemble for multi-npu mode to allow for overriding initial switch config
Summary:
Test ensemble reads the test config from configerator and writes to the disk for hw-agent to consume and initialize.

For some tests, this config is overridden, e.g. add more dsf nodes.

https://www.internalfb.com/code/fbsource/[610682d4ecfbc5d53ff4f511166c30231cadfcb4]/fbcode/fboss/agent/test/agent_hw_tests/AgentVoqSwitchWithMultipleDsfNodesTests.cpp?lines=21-26

Now, sw-switch and hw-agent config are out of sync vis-a-vis their initial config. This leads to crashes in hw-agent  P1796901052 when ports for other dsf nodes are programmed.

Fix
=====
Incorporate the overridden config for hw-agent.

Reviewed By: zechengh09

Differential Revision: D73965000

fbshipit-source-id: 4ca466bdcaf5dc9235a836edd100b129a10ff0ca