Refactor removeFile function: handle EAGAIN
Summary:
This change refactors the removeFile function to handle specific error cases EAGAIN(resource temporarily unavailable).

EAGAIN is not a fatal error and it simply means that the operation could not complete right now, retry should usually solve it
My understaning after reading resources, the kernel is basically telling to back off and try again

We are seeing these error popping up while trying to create warmboot files for hwswitches in Elbert 8DD swicthes.

My hypothesis is that as we are trying multiple parallel unlink calls to files with common prefix, maybe filesystem needs a bit time to update directory metadata (and probably leads to temporary kernel level locks)

We have added a maximum of 3 attempts if we face EAGAIN
Also, I have added 40ms sleep so that we do not try to aggresively and waste cpu cycles

Reviewed By: harshitgulati18

Differential Revision: D74009331

fbshipit-source-id: 6170bf1779039896bf65ab2ef956cfadc3141d8a