Enable Warmboot Feature on Credo Phy ASIC for 0.9.0
Summary:
As titled

This diff stack is making incremental changes for 0.9.0, and most of the lower diffs won't have a test plan since they're not testable on their own.

We'll include comprehensive tests in the topmost diff(s) to make sure everything works as expected for 0.9.0.

Our main priority is to avoid breaking existing code for 0.7.2.

Here we are handling non 0.9.0 case not to trigger gracefulExit of Saiswitch

Reviewed By: srikrishnagopu, birdsoup

Differential Revision: D72523073

fbshipit-source-id: 872fa96a26c154e14b2e5dbafd43a5f622719ead