Remove SrcMac Setting in CredoF104 Warmboot Path
Summary:
Setting the SrcMac during warmboot causes a crash in version 0.9.0, and it's not actually necessary.

For version 0.7.2, this change has no impact since we won't reach that code block on the warmboot path.

As for other SDKs, CredoF104 is the only switch of type SwitchType::PHY, so there are no implications for other switches.

Reviewed By: srikrishnagopu

Differential Revision: D72592460

fbshipit-source-id: 5b30dbc487526e69ddf2273dfdfee5eb6752a25b