Address WB failure with HwVerifyPfcConfigInHwTest
Summary:
The test was failing as we were configuring 200 msec as deadlock detection timer, but
HW was being configured with 198 msec. Now, on WB, we read 198 from HW and that
is different from what SW has which is 200, hence a reconfig is attempted which fails
the "zero SDK write" check. The fix is to make sure we use the same value for config as
 that can be configured in HW.

Reviewed By: maxwindiff

Differential Revision: D74271649

fbshipit-source-id: b0c267ffec32a58e77cb550b46456b93a1d11ece