fan_service: Add a PidLogicBase class to support incremental PID logic
Summary: This diff introduces a new base class PidLogicBase that provides a foundation for implementing different PID logic variant. The primary goal of this change is to facilitate the addition of an incremental pidLogic class, which will be built onto of this class and also PidLogic class will be refactored to inherit from this class.

Reviewed By: somasun, mikechoifb

Differential Revision: D74264027

fbshipit-source-id: b8df2daf8e1204bf2ba626d8908d2a932fc4fffa