Fix bug in preparing state deltas when limit is not exceeded + miscc logging
Summary:
Consider a case where we are adding routes 1,2 .. N and all of these stay
within the ECMP limits. So this should just produce a single delta which is
sent down to HW. To do this our logic was
- Process routes one by one
- If limit is not exceeded, pop the last delta and queue new delta build from
```
deltas.back().newState(), delta.back().newState() + just processed route
```
However this is buggy, if we build deltas like so, we will end up a with a
single delta which just has the delta for route N - 1, route N
Instead since we are in a sense coalescing these deltas, we should have done

```
deltas.back().oldState(), delta.back().newState() + just processed route
```

TODO: there is a coalescing optimization we may be able to do for oveflow
routes as well. Will build that in a separate diff.

Differential Revision:
D74217300

Privacy Context Container: L1125642

fbshipit-source-id: 6db84de2d81f5453b1494ff7b9497f5fe2c4eb78