Ignore config_test.py for running bsp_tests on switch
Summary:
config_test.py is meant to be run internally to test the bsp_tests
code. The validation of actual configs itself happens as part of
fboss/platform/config_lib library. Also, config_test.py does not run as such on
switches as the package assumption made in the file is wrong when copied to a
switch.

Later, we could modify this to not depend on hardcoded list of platform names.

Reviewed By: Scott8440

Differential Revision: D74338629

fbshipit-source-id: 611c6ffa7c7f149c9bb1aa3480537e66d6ef3642