Add test where we are at ECMP limit and update a route to point to new group
Summary:
As titled. This asserts that since ECMP is make before break
we actually do tip over the ECMP limit and have to overflow to back up
ecmp group type.
Once we have the defrag logic, we will rollback this ECMP group back
to primary ecmp type. But for now we do spill over and tests asserts
that.

Differential Revision: D74294260

fbshipit-source-id: 9b2869971c408cc5bc11316b4319dcc1a50339de