Update BcmMultiPathNextHop to include switchingMode
Summary:
Include the switchingMode type as part of the BcmMultiPathNextHopKey. This will allow creating 2 ECMP groups with the same set of nexthops. Helps with cases were routes are migrated between types.

Pass the switchingMode to BcmEcmpEgress to decide the mode. Refer to next diff on how this new param is used in BcmRoute

Reviewed By: msomasundaran

Differential Revision: D74097826

fbshipit-source-id: 3f15b6277678ce244b899259fd0f80b0bf928f18