Janga : Do not turn off the 2nd J3 (J3B) at 18:0.0 by default
Summary:
[FBOSS/Platform] Janga : Do not turn off the 2nd J3 (J3B) at 18:0.0 by default
One of the recent change from the vendor side made Janga platform manager to put the 2nd J3 (J3B, at PCIe 18:0.0) into reset when platform manager loads drivers and initialize the system.

This diff reverts back this behavior - it will keep the reset register value as is, which is 0xff when the system is powered up freshly.

Reviewed By: joancaneus

Differential Revision: D74516544

fbshipit-source-id: 601dd4bccb72ae2797b647457afc53b24307d1cf