Start base state from newState + oldState.fibs()
Summary:
In ECMP resource mgr, we start building new deltas by
traversing the FIB delta and checking for ECMP limit. We still
need to honor any non-FIB changes that came with the original
delta. So in the base delta, start things off from
oldState, newState with Old state FIBs.
Then we traverse FIBs delta and build up new deltas.

Typically we send FIB deltas on their own. With 2
exceptions
- Config application
- Warm boot.

Today we are creating EcmpResourceManager after config
application, so we don't see these update. However
for ZAS, we would need to pass warm boot state through
EcmpResourceManager to reclaim freed up DLB entries.
We could special case this but in any case EcmpResourceManager
should not make assumptions about the kind of delta we get.

Downside of this is that for the case where we have - only
FIBs delta + overflow on the first route. It does create a
extra empty delta in the vector of deltas.

Reviewed By: ashutoshgrewal

Differential Revision:
D74422931

Privacy Context Container: L1125642

fbshipit-source-id: 90794ab8daf7956ae396d967e92913f37055809b