Clean images directory and add utility script
Summary:
Organize the tutorial images that were in the static image directory. Also adds a helper script that will automatically move images to the correct locations and update the markdown files to reflect that.

We needed to define a way to store our images for the OSS documentation site. The most concrete way to do so is to make a directory for each category, then for each file within that category that uses images. Although tedious, this will help us stay organized when adding images to our markdown files. One plus is that we can use scripts and codemods to automatically organize our file structure to ensure we keep the standards in place if somebody were to add images in the same directory as their markdown file instead of the static image directory. It also means we can just let people use the directory they're working in for adding images and we will automatically take care of the organization.

Differential Revision: D74616460