Fix the race condition in SwSwitch with reachability validation
Summary:
In the current model, when a ports ACTIVE state change is noticed in the callback from SDK, we trigger
a switch reachability change fetch from SDK/HW. Now, both the port active state and switch reachability
change reach SwSwitch. In SwSwitch, switch reachability change would trigger a validation of reachability
information which involves cross checking the port active state from state DB with the latest reachability
information received. However, its possible that the state update for port stage change is not yet complete.
So, we need to make sure that we wait for port state update to be complete before doing the validation.
There are 2 options to do this,

1. Trigger reachability validation both after switch reachability update and after port state change update.
2. Trigger switch reachability change in HwSwitch only when the port active state change is available via
state update, which guarantees that SwSwitch has the latest port active state information.

Going with (2) as that avoids the unnecessary transient switch reachability inconsistency detections and
to avoid the additional complexity in (1).

Reviewed By: jasmeetbagga

Differential Revision: D74506270

fbshipit-source-id: 689f76b73219d51f05475e15f2601dc349327e60