Fix build failure and fix a bug
Summary:
Disclamer: I work on thrift and I have no knowledge of fboss codebase.

We (thrift team) switched `MyUnion::Type` to scoped enum, this caused build failure.

However, the existing code is actually checking

```
if (DOMDataUnion::Type::__EMPTY__) {
```

Which would be the equivalent of

```
if (static_cast<int>(DOMDataUnion::Type::__EMPTY__) > 0) {
```

That is probably a bug since this is always false.

It looks like the intention is checking whether the union is empty, so this diff changed to that.

Reviewed By: shiva-menta

Differential Revision: D74633130

fbshipit-source-id: c8f7f5e0346b5c4ce55ff64ec3c75ddd4b2b61be