SerDes extension attributes
Summary:
**TL;DR**
As titled, introducing necessary SAI attributes to program port SerDes settings via SAI API.

So, you might be wondering how KODIAK3 devices are in prod without these changes. So let me explain end-to-end workflow for setting correct SerDes settings. Which will also answer this question. In short (if you don't want read further), we are already programming these via ASIC config in agent config. Now doing the same via SAI APIs.

---------------------------------------------------------------------------

Here are the high level steps:
======================
1. Gather files from Vendor(Cisco)
2. Populate ASIC config section in agent config.
3. Generate platform mapping JSON file.
4. Copy platform mapping JSON content to platform mapping CPP file in fbcode.
5. Implement necessary SAI attributes to program these values via SAI API.

**1. Gather files**:

We receive the following files from Cisco

https://fburl.com/morgan800cc_si_settings
https://fburl.com/morgan800cc_static_mapping
https://fburl.com/morgan800cc_vendor_config

`si_settings` - This file contains SI settings(SerDes parameters) for every lane in the ASIC. For example in KODIAK3 (51.T) chip, we have 512 lanes. So, we will have 512 lines for each speed. Where each line defines the SI settings for a <lane, speed>. This file will be used to create platform mapping JSON.

 `static_mapping` - This file contains lane mapping from ASIC lanes to trasceiver lanes. This is required to by agent to create a port and qsfp_service to program the transceivers. So, you will see one line per lane.

`vendor_config` - This file contains information like device_property, ifg_swap_lists, and serdes_parameters. Content of this file is populated in `asic` config section of `agent` config and passed to SDK as is.

Note that, we do have serdes_parameters section in vendor_config file. Which  defines SI settings for every lane. So, the natural question is, then why do we need `si_settings` file? The answer is, if we want to change/tune these SI setting values, we would end up making changes to `asic` config section of agent config and we will have to push it as disruptive change.

Instead if we set these values as SAI attribute, we can do that via warm boot. And we will have better control as well. More on the later.

**2. Populate ASIC config**
Here we read the vendor_config.json provided by vendor and populate it as is to ASIC config section of agent config. Cisco SDK will consume the ASIC config section and initialize the device.

**3. Generate Platform Mapping**
Platform mapping file is the SoT for several things, to name a few:
i. How many ports are supported.
ii. What speeds are supported.
iii. Which lanes are used for each port.
iv. What are the SerDes settings for each lane.
v. What is the mapping of ASIC iPhy lanes to transceiver lanes.

So, if you notice, to populate all the above information the platform mapping file generator consumes all the three files provided by vendor and generates platform_mapping JSON file in configerator.

So, if you notice, to populate all the above information in the platform mapping, the generator consumes all the three files provided by vendor.

**4. Copy to platform mapping CPP**
The generated platform mapping JSON file in configerator is copied to platform mapping CPP file in fbcode. Agent fetches the required information from platform mapping CPP file and calls SAI APIs to program SDK.

**5. Implement SAI APIs**
When we create ports, we call SAI APIs to set SerDes parameters. This information is read from platform mapping as explained above. In this diff we are implementing SAI attributes for some of the SerDes parameters that were missing.

If we do not program these via SAI APIs, SDK will rely on ASIC config. Which is what is happening for now.

Differential Revision:
D74164676

Privacy Context Container: L1125642

fbshipit-source-id: fe66c17337dc7f244822cdd19b0657b775efdc37